(install.sh) forze unzip to replace files
Previously, if unzip had to replace any files it will try to prompt the
user and mess all the install script, the -o flag tells it to replace.

I also added /dev/tty to its stdin just in case it prompts the user.

Nothing of this should happen anyways with the new tmp location.